Initial Commit
This commit establishes the foundation for the 'carpc' project
repository. It serves as a guide for fetching and building essential
source components. Currently, it includes instructions for
retrieving the following sources:
   - git@github.com:dterletskiy/carpc-builder.git
   - git@github.com:dterletskiy/carpc-tracing.git
   - git@github.com:dterletskiy/carpc-base.git
   - git@github.com:dterletskiy/carpc-tools.git
   - git@github.com:dterletskiy/carpc-osw.git
   - git@github.com:dterletskiy/carpc-runtime.git
   - git@github.com:dterletskiy/carpc-hooks.git
   - git@github.com:dterletskiy/carpc-instrumental.git
   - git@github.com:dterletskiy/carpc-servicebrocker.git
   - git@github.com:dterletskiy/carpc-examples.git
   - git@github.com:dterletskiy/carpc-tutorial.git

A shell script is utilized for automating the fetching and building
procedures, with plans to transition to the 'repo' tool in the
future.

Signed-off-by: Dmitriy Terletskiy <daterletskiy@gmail.com>